+++ apps/web/app/(use-page-wrapper)/settings/(settings-layout)/organizations/(org-admin-only)/attributes/create/page.tsx
@@ -2,7 +2,7 @@ import { _generateMetadata } from "app/_utils";
 
 import OrgAttributesCreatePage from "@calcom/ee/organizations/pages/settings/attributes/attributes-create-view";
 
-import { validateUserHasOrgAdmin } from "../../../_actions/validateUserHasOrgAdmin";
+import { validateUserHasOrgAdmin } from "../../../actions/validateUserHasOrgAdmin";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
