+++ apps/web/app/(use-page-wrapper)/settings/(settings-layout)/organizations/(org-admin-only)/billing/page.tsx
@@ -5,7 +5,7 @@ import SettingsHeader from "@calcom/features/settings/appDir/SettingsHeader";
 
 import BillingView from "~/settings/billing/billing-view";
 
-import { validateUserHasOrgAdmin } from "../../_actions/validateUserHasOrgAdmin";
+import { validateUserHasOrgAdmin } from "../../actions/validateUserHasOrgAdmin";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
