+++ apps/web/app/(use-page-wrapper)/settings/(settings-layout)/organizations/roles/page.tsx
@@ -11,7 +11,7 @@ import { RoleService } from "@calcom/features/pbac/services/role.service";
 import SettingsHeader from "@calcom/features/settings/appDir/SettingsHeader";
 import { prisma } from "@calcom/prisma";
 
-import { validateUserHasOrg } from "../_actions/validateUserHasOrg";
+import { validateUserHasOrg } from "../actions/validateUserHasOrg";
 import { CreateRoleCTA } from "./_components/CreateRoleCta";
 import { RolesList } from "./_components/RolesList";
 import { roleSearchParamsCache } from "./_components/searchParams";
