+++ apps/web/app/(use-page-wrapper)/settings/(settings-layout)/organizations/teams/other/[id]/appearance/page.tsx
@@ -3,7 +3,7 @@ import { _generateMetadata, getTranslate } from "app/_utils";
 import LegacyPage from "@calcom/features/ee/teams/pages/team-appearance-view";
 import SettingsHeader from "@calcom/features/settings/appDir/SettingsHeader";
 
-import { validateUserHasOrg } from "../../../../_actions/validateUserHasOrg";
+import { validateUserHasOrg } from "../../../../actions/validateUserHasOrg";
 
 export const generateMetadata = async ({ params }: { params: Promise<{ id: string }> }) =>
   await _generateMetadata(
